% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CcQueryBuilder.R
\name{buildQuery}
\alias{buildQuery}
\title{Builds SQL code to run analyses according given Cohort Characterization design}
\usage{
buildQuery(cohortCharacterization, cohortTable = "cohort", sessionId,
  cdmSchema, resultsSchema, vocabularySchema = cdmSchema,
  tempSchema = resultsSchema, analysisId)
}
\arguments{
\item{cohortCharacterization}{A string object containing valid JSON that represents CC design}

\item{cohortTable}{The name of table with cohorts}

\item{sessionId}{session identifier using to build temporary tables}

\item{cdmSchema}{the name of schema containing data in CDM format}

\item{resultsSchema}{the name of schema where results would be placed}

\item{vocabularySchema}{the name of schema with vocabularies}

\item{tempSchema}{the name of database temp schema}

\item{analysisId}{analysis identifier}
}
\value{
SQL code in MS Sql Server dialect, if it's required to run analysis on another DBMS
        you have to use \code{\link[SqlRender]{translateSql}} function in the SqlRender package.
}
\description{
Builds SQL code to run analyses according given Cohort Characterization design
}
