% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCohorts.R
\name{cleanupCohortTable}
\alias{cleanupCohortTable}
\title{Drops cohort table}
\usage{
cleanupCohortTable(
  connectionDetails,
  cohortDatabaseSchema,
  cohortTable,
  tempSchema = cohortDatabaseSchema
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cohortDatabaseSchema}{The name of schema where cohort table will be placed}

\item{cohortTable}{The name of cohort table}

\item{tempSchema}{The temp schema if needed}
}
\description{
Drops cohort table
}
