% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCohorts.R
\name{constructCohorts}
\alias{constructCohorts}
\title{Executes given SQLs to construct cohorts}
\usage{
constructCohorts(cohortTableSql, cohortSqls, connectionDetails)
}
\arguments{
\item{cohortTableSql}{SQL to create cohort table}

\item{cohortSqls}{Cohort SQLs}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}
}
\description{
Executes given SQLs to construct cohorts
}
