% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCohorts.R
\name{createCohorts}
\alias{createCohorts}
\title{Constructs cohorts used in CC analysis}
\usage{
createCohorts(pckg, connectionDetails, cdmSchema, vocabularySchema,
  resultSchema)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cohortTableSql}{SQL to create cohort table}

\item{cohortSqls}{Cohort SQLs}
}
\description{
Constructs cohorts used in CC analysis
}
