% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCohorts.R
\name{generateCohortsFromJson}
\alias{generateCohortsFromJson}
\title{Constructs cohorts used in CC analysis}
\usage{
generateCohortsFromJson(
  connectionDetails,
  cdmDatabaseSchema,
  vocabularySchema,
  cohortDatabaseSchema,
  cohortTable,
  tempSchema = cohortDatabaseSchema
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{The schema containing the cdm data}

\item{vocabularySchema}{The schema containing the vocab}

\item{cohortDatabaseSchema}{The schema that will contain the cohortTable}

\item{cohortTable}{The name of the cohortTable that gets created}

\item{tempSchema}{The temp schema if needed}
}
\description{
Constructs cohorts used in CC analysis
}
