% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCohorts.R
\name{getCreateCohortTableSql}
\alias{getCreateCohortTableSql}
\title{Builds SQL to create cohort table}
\usage{
getCreateCohortTableSql(sqlPath, resultSchema, targetTable)
}
\arguments{
\item{sqlPath}{Path to SQL template}

\item{targetTable}{The name of cohort table}

\item{resultsSchema}{The name of schema where cohort table will be placed}
}
\description{
Builds SQL to create cohort table
}
