% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacterization.R
\name{runAnalysis}
\alias{runAnalysis}
\title{Builds query, run analyses and save results to the result folder.}
\usage{
runAnalysis(
  connectionDetails,
  sessionId,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  resultsSchema,
  vocabularySchema,
  tempSchema = resultsSchema,
  jobId = NULL,
  outputFolder = "SkeletonCohortCharacterization",
  customCovariates = NULL,
  createCohorts = T,
  runCharacterization = T,
  saveSql = T,
  saveCvsResults = T,
  viewShiny = T
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{sessionId}{session identifier using to build temporary tables}

\item{cdmDatabaseSchema}{the name of schema containing data in CDM format}

\item{cohortDatabaseSchema}{The name of the schema containing the cohort table}

\item{cohortTable}{The name of table with cohorts}

\item{resultsSchema}{the name of schema where results would be placed}

\item{vocabularySchema}{the name of schema with vocabularies}

\item{tempSchema}{the name of database temp schema}

\item{jobId}{analysis identifier}

\item{outputFolder}{The location to save the results to}

\item{customCovariates}{A list of lists with objects: function (a string) and settings and list of inputs to the function}

\item{createCohorts}{Whether to run the code to create the cohort}

\item{runCharacterization}{Whether to run the code to charcterize the cohorts}

\item{saveSql}{Whether to save the sql used into the outputFolder}

\item{saveCvsResults}{Whether to save the results as csv files}

\item{viewShiny}{Whether to view the results in a shiny app}
}
\description{
Builds query, run analyses and save results to the result folder.
}
