% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results.R
\name{saveResults}
\alias{saveResults}
\title{Save results of analyses from database to CSV files.}
\usage{
saveResults(connectionDetails, cohortCharacterization, analysisId,
  resultsSchema, outputFolder, tresholdLevel = 0.01)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cohortCharacterization}{A string object containing valid JSON that represents CC design}

\item{analysisId}{analysis identifier}

\item{resultsSchema}{the name of schema where results would be placed}

\item{outputFolder}{folder name where results would be saved}

\item{tresholdLevel}{treshold level is used for prevalence analyses to reduce meaningless records having percentage lower than 
treshold level. Default value is 0.01}
}
\description{
Save results of analyses from database to CSV files.
}
