% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCohorts.R
\name{cleanupCohortTable}
\alias{cleanupCohortTable}
\title{Drops cohort table}
\usage{
cleanupCohortTable(connectionDetails, resultSchema, cohortTable)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{resultsSchema}{The name of schema where cohort table will be placed}

\item{targetTable}{The name of cohort table}
}
\description{
Drops cohort table
}
