% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCohorts.R
\name{createCohortTable}
\alias{createCohortTable}
\title{Builds SQL to create cohort table}
\usage{
createCohortTable(
  connectionDetails,
  cohortDatabaseSchema,
  cohortTable,
  oracleTempSchema
)
}
\arguments{
\item{connectionDetails}{The connection details to the OMOP CDM}

\item{cohortDatabaseSchema}{The name of schema where cohort table will be placed}

\item{cohortTable}{The name of cohort table}

\item{oracleTempSchema}{The temp schema}
}
\description{
Builds SQL to create cohort table
}
