% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacterization.R
\name{runAnalysis}
\alias{runAnalysis}
\title{Builds query, run analyses and save results to the result folder.}
\usage{
runAnalysis(connectionDetails, cohortTable = "cohort", sessionId,
  cdmSchema, resultsSchema, vocabularySchema, tempSchema = resultsSchema,
  analysisId, outputFolder = "SkeletonCohortCharacterization")
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cohortTable}{The name of table with cohorts}

\item{sessionId}{session identifier using to build temporary tables}

\item{cdmSchema}{the name of schema containing data in CDM format}

\item{resultsSchema}{the name of schema where results would be placed}

\item{vocabularySchema}{the name of schema with vocabularies}

\item{tempSchema}{the name of database temp schema}

\item{analysisId}{analysis identifier}
}
\description{
Builds query, run analyses and save results to the result folder.
}
