% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\docType{package}
\name{SkeletonCohortCharacterization}
\alias{SkeletonCohortCharacterization}
\alias{SkeletonCohortCharacterization-package}
\title{This package is the skeleton to build own packages with Cohort Characterization analyses. That package is able to run at any site that has access to an observational database in the Common Data Model.}
\description{
This package is the skeleton to build own packages with Cohort Characterization analyses. That package is able to run at any site that has access to an observational database in the Common Data Model.
}
